factor out make_in_between
