+++ scripts/seed_test.py
@@ -1,60 +1,21 @@
 from helpers.inference_spec.sample_spec import SampleSpec
 from helpers.inference_spec.latent_spec import SeedSpec
 from helpers.inference_spec.latents_shape import LatentsShape
-from helpers.inference_spec.cond_spec import ConditionSpec, SingleCondition, MultiCond, WeightedPrompt, CFG, Prompt
+from helpers.inference_spec.cond_spec import SingleCondition, MultiCond, WeightedPrompt, CFG, Prompt
 from helpers.inference_spec.execution_plan_batcher import ExecutionPlanBatcher, BatchSpecGeneric
 from helpers.inference_spec.execution_plan import ExecutionPlan, make_execution_plan
 from helpers.inference_spec.batch_latent_maker import BatchLatentMaker
 from helpers.inference_spec.latent_maker import LatentMaker, MakeLatentsStrategy
 from helpers.inference_spec.latent_maker_seed_strategy import SeedLatentMaker
-from helpers.sample_interpolation.in_between import MakeInbetween, InBetweenParams
+from helpers.sample_interpolation.make_in_between import make_inbetween
 from helpers.sample_interpolation.intersperse_linspace import intersperse_linspace
 from helpers.get_seed import get_seed
 from helpers.device import DeviceLiteral, get_device_type
 from helpers.embed_text_types import Prompts, EmbeddingAndMask
 import torch
 from torch import FloatTensor
-from typing import Iterable, Generator, List, Callable, Tuple, Protocol
+from typing import Iterable, Generator, List
 from itertools import chain, repeat
-
-
-# make_inbetween: MakeInbetween[SingleCondition|MultiCond, MultiCond] = lambda params: MultiCond(
-#   cfg_scale=params.from_.cfg_scale,
-#   weighted_prompts=[*params.from_.weighted_prompts, *params.to.weighted_prompts]
-# )
-class AdjustScale(Protocol):
-  @staticmethod
-  def __call__(scale_nominal: float, quotient: float) -> float: ...
-
-WeightedPromptAndScaleAdjuster = Tuple[WeightedPrompt, AdjustScale]
-
-def make_inbetween(params: InBetweenParams[SingleCondition|MultiCond]) -> MultiCond:
-  assert (params.from_.cfg is None) == (params.to.cfg is None)
-  if params.from_.cfg is not None:
-    assert params.from_.cfg.uncond_prompt.text == params.to.cfg.uncond_prompt.text
-    from_scale: float = params.from_.cfg.scale
-    to_scale: float = params.to.cfg.scale
-    cfg = CFG(scale=from_scale, uncond_prompt=params.from_.cfg.uncond_prompt)
-    cfg_scale_coeff = to_scale / from_scale
-  else:
-    cfg = None
-    cfg_scale_coeff = 1.
-  scale_from: AdjustScale = lambda scale_nominal, quotient: scale_nominal * (1-quotient)
-  scale_to: AdjustScale = lambda scale_nominal, quotient: scale_nominal * quotient * cfg_scale_coeff
-  prompts_and_scale_strategies: Iterable[WeightedPromptAndScaleAdjuster] = chain(
-    zip(params.from_.weighted_cond_prompts, repeat(scale_from)),
-    zip(params.to.weighted_cond_prompts, repeat(scale_to))
-  )
-
-  return MultiCond(
-    cfg=cfg,
-    weighted_cond_prompts=[
-      WeightedPrompt(
-        wp.prompt,
-        scale(wp.weight, params.quotient)
-      ) for wp, scale in prompts_and_scale_strategies
-    ]
-  )
 
 cond_keyframes: List[SingleCondition|MultiCond] = [SingleCondition(
   cfg=CFG(scale=7.5, uncond_prompt=Prompt(text='')),
